created a credentials package with keys for use with APIs
